Generate only merged coverage
MSP-12019

Coverage is 100% only if the coverage for spec and cucumber are merged,
so only generate coverage reports if both are run.  This can be done by
running the `coverage` task, which is now the default task.  Coverage
merging is also used for coveralls.io.  CodeClimate coverage is disabled
again.