Check merged results in coverage for minimum coverage and drop
MSP-12148

'simplecov/defaults' at_exit only works on unmerged results.  Only
merged results exist in the process when the 'coverage' task is run
(because the 'cucumber' and 'spec' tasks shell out to run cucumber and
rspec), so the at_exit behavior had to be copied and adapted to work on
merged results.