Add :nocov: around file cleanup code in specs
MSP-12083

File cleanup code is only exercised if the specs fail and they're not
failing, so the code won't run normally and will show as uncovered.
Coverage for this cleanup code is not important enough to fake an
erroring testing in cucumber because if the cleanup is broke then that
will signal that the tests were broke too.