Add :content? flag to disable writing template
MSP-12300

A few specs expect the content not to be written to disk, so I added a
`content?` transient to the :metasploit_cache_module_ancestor` trait to
disable the `after(:build)`.  Additionally, if the default `content?:
true` is used, but `module_ancestor.real_pathname` is `nil`, then an
`ArgumentError` is raised explaining to use `content?: false`.