Validate compatible platforms between stage and stager
MSP-12430

Unlike `#compatible_architectures`, calculating if the platforms are
compatible is more complex because platforms are a nested set.  Like
validating compatible architectures, I wanted to be able to get the list
of compatible platforms, so I needed to decide whether the the staged
supported the stage, stager, or both (because they are the same)'s
platform.  In general, whichever platform is the subset should chosen,
so between 'Windows` and 'Windows 95', 'Windows 95' should be chosen.
This needs to be done for if either the stage or the stager has the
subset, so this means taking the UNION of the subset calculation where
the position of the stager and stage platforms are reversed.  Since this
would involve reused queries, I used Common Table Expressions (CTE),
which are supported in AREL, but not ActiveRecord::Relation, to name the
`payload_stager_platforms` and `payload_stage_platforms`.