Ensure Auxiliary::Instance associations are built before writing contents
MSP-13028

With
:metasploit_cache_auxiliary_instance_auxiliary_class_ancestor_contents
as a trait, it's callback to write the contents was run before the
callback in the :metasploit_cache_auxiliary_instance factory proper to
make the has_many associations.  This left the generated content with
empty collections for actions, authors and license even though the
factory had actions, contributions, and licensable_licenses.