Better Exploit::Instance#stance syncing
MSP-12444

Like #license, #stance transforms from String to Array<String> if more
than one mixin has a 'Stance'.  This happens in exploits in
metasploit-framework that mixin servers, which are passive, and clients,
which are aggressive.  If #stance is Array<String> and it contains
'aggressive', then Metasploit::Cache::Exploit::Instance#stance is
'aggressive'.

'aggressive' => 'aggressive'
'passive' => 'passive'
['aggressive', 'passive'] => 'aggressive'