Log and validate instead of saving again in batched_save
MSP-12444

Sqlite3 and Postgresql to a lesser extent doesn't always show the same
error when save is called twice, including odd cases where no values are
used for a an INSERT and DEFAULT is used instead.  Because the errors
were unhelpful and the whole point of the second, non-batched save is to
get better errors, move to logging the original exception and then just
validating.  This also protects from issues that will cause an error on
the second save, such as non-unique rows that are both in-memory at the
time of save and so only caught by the database constraints.