Metasploit Module classes must have framework attribute
MSP-12444

A few Metasploit Modules from metasploit-framework access `#framework`
from inside `#initialize`.  The instance `#framework` delegates to the
class `framework`, so the class must support setting `framework` with
`framework=` so `#initialize` can see if without passing it as an
argument to `new`.

NOTE: the framework SHOULD be passed to `new` in the future, but that
requires changing metasploit-framework, which is prohibitted prior to
integration.