Payload::Staged::Class::Load
MSP-13082

To load a staged class, 3 Modules are required: the handler module, the
payload stage Module, and the payload stager Module.  Like the other
loads, an ephemeral_cache_by_source is attached, but also an
ancestor_by_source because there needs to be an easy way to look up the
stage and stager modules to access their ephemeral_cache_by_source.