Payload::Direct -> Payload::Unhandled
MSP-13201

The name Payload::Direct is slightly misleading in that it's not a
direct class from the contents of the ancestor, but made by only mixing
the module from contents of the ancestor into Msf::Payloads.
Additionally, the important distinction between that type of payloads
classes and those used in msfconsole is that the msfconsole have the
handler mixed in.  To reflect the importance of the handler mixin,
Payload::Direct is renamed Payload::Unhandled.