metasploit-cache command line tool
MSP-12459

`metasploit-cache load MODULE_PATH` loads the modules on a MODULE_PATH
into the cache.  `metasploit-cache use FULL_NAME` then attempts to
instantiation the Metasploit Module with the FULL_NAME from the built
cache.  Actualy testing in rspec had to be limited to just encoders and
nops because the other module types takes too long to test because each
example takes ~5 seconds with 3 seconds devoted to the `require
'metasploit/framework'`.  The tests for auxiliary, exploit, and post
can't be enabled until the loading time of 'metasploit/framework' is
improved.